requirement.txt
requests>=2.26.0
beautifulsoup4>=4.10.0
spacy>=3.3.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
slack-sdk>=3.15.0
python-telegram-bot>=13.7
fastapi>=0.70.0
uvicorn>=0.15.0